requirements.txt
streamlit
soundfile
openai
requests
python-dotenv
streamlit-mic-recorder
pydub